[FE] Move OnlyForDefaultLanguageFeatureDisabled to core.language.version-settings
... and allow providing more than one feature
